- Removed findCloudComponents, findConnectedComponents, computeBasicStats, remove findCloudElements, checkCriteria and checkComponents Overlap from MCCOps - Moved reduceResolution, reduceRectangleResolution, and LabelConnectedComponents implementations to AbstractTensor - Renamed LabelConnectedComponents to LabelComponents
