+++ src/main/scala/org/dia/algorithms/mcc/MCCOps.scala
@@ -34,147 +34,146 @@ import org.dia.tensors.AbstractTensor
  */
 object MCCOps {
 
-  val BACKGROUND = 0.0
   val logger = org.slf4j.LoggerFactory.getLogger(this.getClass)
-  /**
-   * Reduces the resolution of the tensor into square blocks
-   * by taking the average.
-   *
-   * @param tensor original tensor
-   * @param blockSize block size used for aggregation
-   * @param invalid value used to signify invalid value
-   * @return aggregated tensor obtained by taking the average of
-   * values within blocks ignoring invalid values
-   */
-  def reduceResolution(tensor: AbstractTensor, blockSize: Int, invalid: Double = Double.NaN): AbstractTensor = {
-    val largeArray = tensor
-    val numRows = largeArray.rows()
-    val numCols = largeArray.cols()
-    val reducedMatrix = tensor.zeros(numRows / blockSize, numCols / blockSize)
+  // /**
+  //  * Reduces the resolution of the tensor into square blocks
+  //  * by taking the average.
+  //  *
+  //  * @param tensor original tensor
+  //  * @param blockSize block size used for aggregation
+  //  * @param invalid value used to signify invalid value
+  //  * @return aggregated tensor obtained by taking the average of
+  //  * values within blocks ignoring invalid values
+  //  */
+  // def reduceResolution(tensor: AbstractTensor, blockSize: Int, invalid: Double = Double.NaN): AbstractTensor = {
+  //   val largeArray = tensor
+  //   val numRows = largeArray.rows()
+  //   val numCols = largeArray.cols()
+  //   val reducedMatrix = tensor.zeros(numRows / blockSize, numCols / blockSize)
 
-    for (row <- 0 until reducedMatrix.rows) {
-      for (col <- 0 until reducedMatrix.cols) {
-        val rowRange = (row * blockSize) -> ((row + 1) * blockSize)
-        val columnRange = (col * blockSize) -> ((col + 1) * blockSize)
-        val block = tensor(rowRange, columnRange).copy
-        val numValid = block.data.count(_ != invalid)
-        val avg = if (numValid > 0) block.cumsum / numValid else 0.0
-        reducedMatrix.put(avg, row, col)
-      }
-    }
-    reducedMatrix
-  }
+  //   for (row <- 0 until reducedMatrix.rows) {
+  //     for (col <- 0 until reducedMatrix.cols) {
+  //       val rowRange = (row * blockSize) -> ((row + 1) * blockSize)
+  //       val columnRange = (col * blockSize) -> ((col + 1) * blockSize)
+  //       val block = tensor(rowRange, columnRange).copy
+  //       val numValid = block.data.count(_ != invalid)
+  //       val avg = if (numValid > 0) block.cumsum / numValid else 0.0
+  //       reducedMatrix.put(avg, row, col)
+  //     }
+  //   }
+  //   reducedMatrix
+  // }
 
-  /**
-   * Reduces the resolution of the tensor into rectangle blocks
-   * by taking the average.
-   *
-   * Similar to above method reduceResolution.
-   */
-  def reduceRectangleResolution(
-      tensor: AbstractTensor,
-      rowSize: Int,
-      colSize: Int,
-      invalid: Double = Double.NaN): AbstractTensor = {
+  // /**
+  //  * Reduces the resolution of the tensor into rectangle blocks
+  //  * by taking the average.
+  //  *
+  //  * Similar to above method reduceResolution.
+  //  */
+  // def reduceRectangleResolution(
+  //     tensor: AbstractTensor,
+  //     rowSize: Int,
+  //     colSize: Int,
+  //     invalid: Double = Double.NaN): AbstractTensor = {
 
-    val largeArray = tensor
-    val numRows = largeArray.rows()
-    val numCols = largeArray.cols()
-    val reducedMatrix = tensor.zeros(numRows / rowSize, numCols / colSize)
+  //   val largeArray = tensor
+  //   val numRows = largeArray.rows()
+  //   val numCols = largeArray.cols()
+  //   val reducedMatrix = tensor.zeros(numRows / rowSize, numCols / colSize)
 
-    for (row <- 0 until reducedMatrix.rows) {
-      for (col <- 0 until reducedMatrix.cols) {
-        val rowRange = (row * rowSize) -> ((row + 1) * rowSize)
-        val columnRange = (col * colSize) -> ((col + 1) * colSize)
-        val block = tensor(rowRange, columnRange).copy
-        val numValid = block.data.count(_ != invalid)
-        val avg = if (numValid > 0) block.cumsum / numValid else 0.0
-        reducedMatrix.put(avg, row, col)
-      }
-    }
-    reducedMatrix
-  }
+  //   for (row <- 0 until reducedMatrix.rows) {
+  //     for (col <- 0 until reducedMatrix.cols) {
+  //       val rowRange = (row * rowSize) -> ((row + 1) * rowSize)
+  //       val columnRange = (col * colSize) -> ((col + 1) * colSize)
+  //       val block = tensor(rowRange, columnRange).copy
+  //       val numValid = block.data.count(_ != invalid)
+  //       val avg = if (numValid > 0) block.cumsum / numValid else 0.0
+  //       reducedMatrix.put(avg, row, col)
+  //     }
+  //   }
+  //   reducedMatrix
+  // }
 
-  /**
-   * Computes connected component labels of tensor.
-   *
-   * Note that for garbage collection purposes we use one ArrayStack.
-   * We push in row/col tuples two ints at a time, and pop two ints at a time.
-   * Also note that the components are labeled starting at 1.
-   *
-   * This is just doing a depth-first-search (DFS) over the tensor cells.
-   * (The tensor cells are a graph with two cells being connected iff
-   * they are next to each other horizontally or vertically (but not diagonally).)
-   *
-   * @param tensor the tensor
-   * @return a single tensor with the associated component numbers,
-   * together with the total number of components.
-   */
-  def labelConnectedComponents(tensor: AbstractTensor): (AbstractTensor, Int) = {
-    val fourVector = List((1, 0), (-1, 0), (0, 1), (0, -1))
-    val rows = tensor.rows()
-    val cols = tensor.cols()
-    val labels = tensor.zeros(tensor.shape: _*)
-    var label = 1
-    val stack = new util.ArrayDeque[Int](tensor.rows + tensor.cols * 10)
+  // /**
+  //  * Computes connected component labels of tensor.
+  //  *
+  //  * Note that for garbage collection purposes we use one ArrayStack.
+  //  * We push in row/col tuples two ints at a time, and pop two ints at a time.
+  //  * Also note that the components are labeled starting at 1.
+  //  *
+  //  * This is just doing a depth-first-search (DFS) over the tensor cells.
+  //  * (The tensor cells are a graph with two cells being connected iff
+  //  * they are next to each other horizontally or vertically (but not diagonally).)
+  //  *
+  //  * @param tensor the tensor
+  //  * @return a single tensor with the associated component numbers,
+  //  * together with the total number of components.
+  //  */
+  // def labelConnectedComponents(tensor: AbstractTensor): (AbstractTensor, Int) = {
+  //   val fourVector = List((1, 0), (-1, 0), (0, 1), (0, -1))
+  //   val rows = tensor.rows()
+  //   val cols = tensor.cols()
+  //   val labels = tensor.zeros(tensor.shape: _*)
+  //   var label = 1
+  //   val stack = new util.ArrayDeque[Int](tensor.rows + tensor.cols * 10)
 
-    /**
-     * Whether we already assigned a component label to the coordinate.
-     *
-     * (row,col) is already labeled if
-     * (i) it's out of bounds OR
-     * (ii) its value is BACKGROUND (=0) OR
-     * (iii) it has already been assigned to a component,
-     * i.e. labels(row,col) is not BACKGROUND (=0)
-     *
-     * @param row the row to check
-     * @param col the column to check
-     * @return whether (row,col) is already labeled
-     */
-    def isLabeled(row: Int, col: Int): Boolean = {
-      if (row < 0 || col < 0 || row >= rows || col >= cols) return true
-      tensor(row, col) == BACKGROUND || labels(row, col) != BACKGROUND
-    }
+  //   /**
+  //    * Whether we already assigned a component label to the coordinate.
+  //    *
+  //    * (row,col) is already labeled if
+  //    * (i) it's out of bounds OR
+  //    * (ii) its value is BACKGROUND (=0) OR
+  //    * (iii) it has already been assigned to a component,
+  //    * i.e. labels(row,col) is not BACKGROUND (=0)
+  //    *
+  //    * @param row the row to check
+  //    * @param col the column to check
+  //    * @return whether (row,col) is already labeled
+  //    */
+  //   def isLabeled(row: Int, col: Int): Boolean = {
+  //     if (row < 0 || col < 0 || row >= rows || col >= cols) return true
+  //     tensor(row, col) == BACKGROUND || labels(row, col) != BACKGROUND
+  //   }
 
-    /**
-     * Pushes unlabeled neighbors onto the stack.
-     *
-     * @param currentLabel the component label being added
-     */
-    def dfs(currentLabel: Int): Unit = {
-      while (!stack.isEmpty) {
-        /**
-         *  Note that when popping, we pop col, then row.
-         *  When pushing, we push row, then col.
-         */
-        val col = stack.pop
-        val row = stack.pop
-        labels.put(currentLabel, row, col)
-        val neighbors = fourVector.map(p => (p._1 + row, p._2 + col))
-        for (neighbor <- neighbors) {
-          if (!isLabeled(neighbor._1, neighbor._2)) {
-            val row = neighbor._1
-            val col = neighbor._2
-            stack.push(row)
-            stack.push(col)
-          }
-        }
-      }
-    }
+  //   /**
+  //    * Pushes unlabeled neighbors onto the stack.
+  //    *
+  //    * @param currentLabel the component label being added
+  //    */
+  //   def dfs(currentLabel: Int): Unit = {
+  //     while (!stack.isEmpty) {
+  //       /**
+  //        *  Note that when popping, we pop col, then row.
+  //        *  When pushing, we push row, then col.
+  //        */
+  //       val col = stack.pop
+  //       val row = stack.pop
+  //       labels.put(currentLabel, row, col)
+  //       val neighbors = fourVector.map(p => (p._1 + row, p._2 + col))
+  //       for (neighbor <- neighbors) {
+  //         if (!isLabeled(neighbor._1, neighbor._2)) {
+  //           val row = neighbor._1
+  //           val col = neighbor._2
+  //           stack.push(row)
+  //           stack.push(col)
+  //         }
+  //       }
+  //     }
+  //   }
 
-    /** Main loop */
-    for (row <- 0 until rows) {
-      for (col <- 0 until cols) {
-        if (!isLabeled(row, col)) {
-          stack.push(row)
-          stack.push(col)
-          dfs(label)
-          label += 1
-        }
-      }
-    }
-    (labels, label - 1)
-  }
+  //   /** Main loop */
+  //   for (row <- 0 until rows) {
+  //     for (col <- 0 until cols) {
+  //       if (!isLabeled(row, col)) {
+  //         stack.push(row)
+  //         stack.push(col)
+  //         dfs(label)
+  //         label += 1
+  //       }
+  //     }
+  //   }
+  //   (labels, label - 1)
+  // }
 
   /**
    * Whether two components overlap.
